+++ src/command_modules/azure-cli-botservice/azure/cli/command_modules/botservice/_params.py
@@ -16,7 +16,7 @@ name_arg_type = CLIArgumentType(metavar='NAME', configured_default='botname', id
 # supported_languages will be use with get_enum_type after the 'Node' value is completely removed from `az bot`
 # In custom.py we're still supporting 'Node' in __language_validator()
 SUPPORTED_LANGUAGES = ['Csharp', 'Javascript']
-UPCOMING_LANGUAGES = list(SUPPORTED_LANGUAGES).append('Typescript')
+UPCOMING_LANGUAGES = ['Csharp', 'Javascript', 'Typescript']
 SUPPORTED_APP_INSIGHTS_REGIONS = [
     'Australia East',
     'Canada Central',
@@ -33,8 +33,6 @@ SUPPORTED_APP_INSIGHTS_REGIONS = [
     'UK South',
     'West Europe',
     'West US 2']
-SUPPORTED_LANGUAGES_MSG = '  Allowed values: Csharp, Javascript.'
-UPCOMING_SUPPORTED_LANGUAGES_MSG = '  Allowed values: Csharp, Javascript, Typescript.'
 SUPPORTED_SKUS = ['F0', 'S1']
 
 
@@ -44,8 +42,8 @@ def load_arguments(self, _):
         c.argument('code_dir', options_list=['--code-dir'], help='The directory to place the generated deployment '
                                                                  'files in. Defaults to the current directory the '
                                                                  'command is called from.')
-        c.argument('language', options_list=['--lang'], help='The language or runtime of the bot.{0}'
-                   .format(UPCOMING_SUPPORTED_LANGUAGES_MSG))
+        c.argument('language', options_list=['--lang'], help='The language or runtime of the bot.',
+                   arg_type=get_enum_type(UPCOMING_LANGUAGES))
         c.argument('proj_file_path', help='The path to the .csproj file relative to --code-dir.')
 
     with self.argument_context('bot') as c:
@@ -56,19 +54,20 @@ def load_arguments(self, _):
                    arg_type=name_arg_type)
 
     with self.argument_context('bot create') as c:
-        c.argument('sku_name', options_list=['--sku'], arg_type=get_enum_type(SUPPORTED_SKUS), help='The Sku of the bot.', arg_group='Registration bot Specific')
+        c.argument('sku_name', options_list=['--sku'], arg_type=get_enum_type(SUPPORTED_SKUS), help='The Sku of the bot.', arg_group='Registration Bot Specific')
         c.argument('kind', options_list=['--kind', '-k'], arg_type=get_enum_type(['registration', 'function', 'webapp']), help='The kind of the bot.')
-        c.argument('display_name', help='The display name of the bot. If not specified, defaults to the name of the bot.', arg_group='Registration bot Specific')
-        c.argument('description', options_list=['--description', '-d'], help='The description of the bot.', arg_group='Registration bot Specific')
-        c.argument('endpoint', options_list=['-e', '--endpoint'], help='The messaging endpoint of the bot.', arg_group='Registration bot Specific')
+        c.argument('display_name', help='The display name of the bot. If not specified, defaults to the name of the bot.', arg_group='Registration Bot Specific')
+        c.argument('description', options_list=['--description', '-d'], help='The description of the bot.', arg_group='Registration Bot Specific')
+        c.argument('endpoint', options_list=['-e', '--endpoint'], help='The messaging endpoint of the bot.', arg_group='Registration Bot Specific')
         c.argument('msa_app_id', options_list=['--appid'], help='The Microsoft account ID (MSA ID) to be used with the bot.')
-        c.argument('password', options_list=['-p', '--password'], help='The Microsoft account (MSA) password for the bot.')
-        c.argument('storageAccountName', options_list=['-s', '--storage'], help='WARNING: Not used in V4 bot creation. Storage account name to be used with the bot. If not provided, a new account will be created.', arg_group='Web/Function bot Specific')
+        c.argument('password', options_list=['-p', '--password'], help='The Microsoft account (MSA) password for the bot. Used to authorize messages being sent to the bot.')
+        c.argument('storageAccountName', options_list=['-s', '--storage'], help='WARNING: Not used in V4 bot creation. Storage account name to be used with the bot. If not provided, a new account will be created.', arg_group='Web/Function Bot Specific')
         c.argument('tags', arg_type=tags_type)
-        c.argument('language', help='The language to be used to create the bot.{0}'.format(SUPPORTED_LANGUAGES_MSG), options_list=['--lang'], arg_group='Web/Function bot Specific')
-        c.argument('appInsightsLocation', help='WARNING: Not used in V4 bot creation. The location for the app insights to be used with the bot.  Default: South Central US.', options_list=['--insights-location'], arg_group='Web/Function bot Specific',
+        c.argument('language', options_list=['--lang'], arg_type=get_enum_type(SUPPORTED_LANGUAGES), help='The language to be used to create the bot.', arg_group='Web/Function Bot Specific')
+        c.argument('appInsightsLocation', help='WARNING: Not used in V4 bot creation. The location for the app insights to be used with the bot.  Default: South Central US.', options_list=['--insights-location'], arg_group='Web/Function Bot Specific',
                    arg_type=get_enum_type(SUPPORTED_APP_INSIGHTS_REGIONS))
-        c.argument('version', options_list=['-v', '--version'], help='The Microsoft Bot Builder SDK version to be used to create the bot', arg_type=get_enum_type(['v3', 'v4']), arg_group='Web/Function bot Specific')
+        c.argument('version', options_list=['-v', '--version'], help='The Microsoft Bot Builder SDK version to be used to create the bot', arg_type=get_enum_type(['v3', 'v4']), arg_group='Web/Function Bot Specific')
+        c.argument('deploy_echo', options_list=['--echo'], arg_type=get_three_state_flag(), help='Deploy an Echo Bot template to the newly created v4 Web App Bot.', arg_group='V4 Bot Templates')
 
     with self.argument_context('bot publish') as c:
         c.argument('code_dir', options_list=['--code-dir'], help='The directory to upload bot code from.')